+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- Add oauth2 and OIDC authentication backends for SSO with Open edX
+
 ## [2.0.0-beta.8] - 2020-06-17
 
 ### Added